requirements.txt
annotated-doc==0.0.4
bitmath==2.0.1
build==1.3.0
cfgv==3.5.0
click==8.3.3
distlib==0.4.0
filelock==3.29.0
identify==2.6.19
iniconfig==2.3.0
librt==0.9.0
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy==1.20.2
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.2
pathspec==1.1.1
platformdirs==4.9.6
pluggy==1.6.0
pre-commit==4.6.0
Pygments==2.20.0
pyproject_hooks==1.2.0
pytest==9.0.3
python-discovery==1.2.2
PyYAML==6.0.3
rich==15.0.0
ruff==0.15.12
setuptools==80.9.0
shellingham==1.5.4
textual==8.2.5
twine==6.1.0
typer==0.25.1
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
uc-micro-py==2.0.0
virtualenv==21.3.0
wheel==0.45.1
